room generator works

+++ src/mega/trigger/spawn_trigger_cbs.c
@@ -44,7 +44,7 @@ void trigger_cb(u8 self){
 		//we execute our fading stuff
 		int i;
 		for (i = 0; i < *battler_cnt; i++){
-			if (battlers[i].trigger && !is_opponent(i)){
+			if (battlers[i].trigger && !is_opponent((u8)i)){
 				//we find the palette 
 				u8 oam_id = battler_oams[i];
 				u8 pal = (u8)(((oams[oam_id].final_oam.attr2 >> 12) & 15)+16);
